LockedApp crate which is independent of spin-core
