+++ Cargo.toml
@@ -54,6 +54,7 @@ spin-doctor = { path = "crates/doctor" }
 spin-http = { path = "crates/http" }
 spin-trigger-http = { path = "crates/trigger-http" }
 spin-loader = { path = "crates/loader" }
+spin-locked-app = { path = "crates/locked-app" }
 spin-manifest = { path = "crates/manifest" }
 spin-oci = { path = "crates/oci" }
 spin-plugins = { path = "crates/plugins" }
