+++ crates/loader/Cargo.toml
@@ -24,7 +24,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sha2 = "0.10.8"
 shellexpand = "3.1"
-spin-app = { path = "../app" }
+spin-locked-app = { path = "../locked-app" }
 spin-common = { path = "../common" }
 spin-config = { path = "../config" }
 spin-manifest = { path = "../manifest" }
