+++ crates/locked-app/src/values.rs
